Refactoring: clean render & fix chest
